Trait xml_data::parser::ElementParser

pub trait ElementParser: Sized {
    fn parse_element_state<E: ElementState>(self, state: &mut E) -> Result<()>;
}

A parser that is ready to parse exactly one element (and nested data).

Required methods

fn parse_element_state<E: ElementState>(self, state: &mut E) -> Result<()>

Start parsing an element with the prepared state

A parser will call the various ElementState to parse the element.

Users of this method will create the state using ElementState::parse_element_start and produce the final result using ElementState::parse_element_finish after calling this method.