Trait xml_data::parser::Value

pub trait Value<T> {
    fn parse_value(text: Cow<str>) -> Result<T>;
}

Trait to parse attributes and inner text

This is implemented my "marker" types to decide how to parse a type (the same type can be parsed differently depending on the marker type)

Required methods

fn parse_value(text: Cow<str>) -> Result<T>

Serialize value to text

Loading content...