Trait xml_data::serializer::Value

pub trait Value<T> {
    fn serialize_value(data: &T) -> Result<Cow<str>>;
}

Trait to serialize attributes and inner text

This is implemented my "marker" types to decide how to serialize a type (the same type can be serialized differently depending on the marker type)

Required methods

fn serialize_value(data: &T) -> Result<Cow<str>>

Serialize value to text

Loading content...